requirements.txt
streamlit
langchain
python-dotenv
pypdf
chromadb
replicate